budget: colocate Rule, NormalizationRule, WeeklyAggregate (#563)
Final three entities consolidated to entities/. Slims
budget-seed-data.d.ts to a thin module declaration that re-exports
SeedX types from entity files (was 129 lines of duplicated literal
unions; now 29). converters.ts becomes a re-export shim, and the
cross-cutting orchestration files (firestore.ts, upload.ts,
export.ts, vite-plugin-seed-data.ts) now delegate read-side parsing
to per-entity functions.

Closes the per-entity colocation refactor: each budget entity has
one source-of-truth file under budget/src/entities/ containing all
of its representations and parsers/converters/serializers.

Refs #563.
