+++ budget/src/budget-seed-data.d.ts
@@ -1,116 +1,16 @@
 // Type declarations for the virtual module generated by vite-plugin-seed-data.ts at build time.
-// These types correspond to the IDB interfaces with ms-based timestamps instead of Firestore Timestamps.
-// Fields validated at serialization time are non-optional here, unlike IDB where older records may lack them.
+// Per-entity Seed* types are defined in their entity files; this module re-exports them
+// under the names callers import from "virtual:budget-seed-data".
 declare module "virtual:budget-seed-data" {
-  export interface SeedTransaction {
-    readonly id: string;
-    readonly institution: string;
-    readonly account: string;
-    readonly description: string;
-    readonly amount: number;
-    readonly note: string;
-    readonly category: string;
-    readonly reimbursement: number;
-    readonly budget: string | null;
-    readonly timestampMs: number | null;
-    readonly statementId: string | null;
-    readonly statementItemId: string | null;
-    readonly normalizedId: string | null;
-    readonly normalizedPrimary: boolean;
-    readonly normalizedDescription: string | null;
-    readonly virtual: boolean;
-  }
-
-  export interface SeedBudgetOverride {
-    readonly dateMs: number;
-    readonly balance: number;
-  }
-
-  export interface SeedBudget {
-    readonly id: string;
-    readonly name: string;
-    readonly allowance: number;
-    readonly allowancePeriod: "weekly" | "monthly" | "quarterly";
-    readonly rollover: "none" | "debt" | "balance";
-    readonly overrides: SeedBudgetOverride[];
-  }
-
-  export interface SeedBudgetPeriod {
-    readonly id: string;
-    readonly budgetId: string;
-    readonly periodStartMs: number;
-    readonly periodEndMs: number;
-    readonly total: number;
-    readonly count: number;
-    readonly categoryBreakdown: Record<string, number>;
-  }
-
-  export interface SeedRule {
-    readonly id: string;
-    readonly type: "categorization" | "budget_assignment";
-    readonly pattern: string;
-    readonly target: string;
-    readonly priority: number;
-    readonly institution: string | null;
-    readonly account: string | null;
-    readonly minAmount: number | null;
-    readonly maxAmount: number | null;
-    readonly excludeCategory: string | null;
-    readonly matchCategory: string | null;
-  }
-
-  export interface SeedNormalizationRule {
-    readonly id: string;
-    readonly pattern: string;
-    readonly patternType: string | null;
-    readonly canonicalDescription: string;
-    readonly dateWindowDays: number;
-    readonly institution: string | null;
-    readonly account: string | null;
-    readonly priority: number;
-  }
-
-  export interface SeedStatement {
-    readonly id: string;
-    readonly statementId: string;
-    readonly institution: string;
-    readonly account: string;
-    readonly balance: number;
-    readonly period: string;
-    readonly balanceDate: string | null;
-    readonly lastTransactionDateMs: number | null;
-    readonly virtual: boolean;
-  }
-
-  export interface SeedStatementItem {
-    readonly id: string;
-    readonly statementItemId: string;
-    readonly statementId: string;
-    readonly institution: string;
-    readonly account: string;
-    readonly period: string;
-    readonly amount: number;
-    readonly timestampMs: number;
-    readonly description: string;
-    readonly fitid: string;
-  }
-
-  export interface SeedReconciliationNote {
-    readonly id: string;
-    readonly entityType: "transaction" | "statementItem";
-    readonly entityId: string;
-    readonly classification: "timing" | "missing_entry" | "discrepancy";
-    readonly note: string;
-    readonly updatedAtMs: number;
-    readonly updatedBy: string;
-  }
-
-  export interface SeedWeeklyAggregate {
-    readonly id: string;
-    readonly weekStartMs: number;
-    readonly creditTotal: number;
-    readonly unbudgetedTotal: number;
-  }
+  export type { SeedTransaction } from "./entities/transaction";
+  export type { SeedBudgetOverride, SeedBudget } from "./entities/budget";
+  export type { SeedBudgetPeriod } from "./entities/budget-period";
+  export type { SeedRule } from "./entities/rule";
+  export type { SeedNormalizationRule } from "./entities/normalization-rule";
+  export type { SeedStatement } from "./entities/statement";
+  export type { SeedStatementItem } from "./entities/statement-item";
+  export type { SeedReconciliationNote } from "./entities/reconciliation-note";
+  export type { SeedWeeklyAggregate } from "./entities/weekly-aggregate";
 
   export interface SeedData {
     readonly transactions: SeedTransaction[];
